Fix tests for CardWrapper and extract some functions to cardUtils.ts
