+++ src/app/content/highlights/components/cardUtils.ts
@@ -1,0 +1,40 @@
+import { Highlight } from '@openstax/highlighter';
+import { HTMLElement } from '@openstax/types/lib.dom';
+import { findElementSelfOrParent } from '../../../domUtils';
+import { assertWindow } from '../../../utils';
+
+export const getHighlightOffset = (container: HTMLElement | undefined, highlight: Highlight) => {
+  if (!container || !highlight.range || !highlight.range.getBoundingClientRect) {
+    return;
+  }
+
+  const {top, bottom } = highlight.range.getBoundingClientRect();
+
+  const offsetParent = container.offsetParent && findElementSelfOrParent(container.offsetParent);
+  const parentOffset = offsetParent ? offsetParent.offsetTop : 0;
+  const scrollOffset = assertWindow().scrollY;
+
+  return {
+    bottom: bottom - parentOffset + scrollOffset,
+    top: top - parentOffset + scrollOffset,
+  };
+};
+
+export const getHighlightTopOffset = (container: HTMLElement | undefined, highlight: Highlight): number | undefined => {
+  const offset = getHighlightOffset(container, highlight);
+
+  if (offset) {
+    return offset.top;
+  }
+};
+
+export const getHighlightBottomOffset = (
+  container: HTMLElement | undefined,
+  highlight: Highlight
+): number | undefined => {
+  const offset = getHighlightOffset(container, highlight);
+
+  if (offset) {
+    return offset.bottom;
+  }
+};
